pyproject.toml
[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"

[project]
name = "hla_algorithm"
description = ''
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = []
authors = [
  { name = "Rosemary McCloskey", email = "retired@bccfe.ca" },
  { name = "David Rickett", email = "drickett@bccfe.ca" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]

dynamic = ["version"]
dependencies = [
  "biopython>=1.85",
  "numpy>=2.2.3",
  "pydantic>=2.10.6",
  "pydantic-numpy>=8.0.1",
  "pyyaml>=6.0.2",
  "requests>=2.32.3",
  "typer>=0.15.2",
  "types-pyyaml>=6.0.12.20250516",
  "types-requests>=2.32.4.20250611",
]

[dependency-groups]
dev = [
  "mypy>=1.15.0",
  "mypy-extensions>=1.0.0",
  "pandas>=2.2.3",
  "pandas-stubs>=2.2.3.241126",
  "pre-commit>=4.1.0",
  "pytest>=8.3.5",
  "pytest-cov>=6.0.0",
  "pytest-html>=4.1.1",
  "pytest-mock>=3.14.0",
  "pytest-xdist>=3.6.1",
  "ruff>=0.9.9",
  "uv-dynamic-versioning>=0.8.2",
]

[project.urls]
Documentation = "https://github.com/cfe-lab/hla_algorithm/blob/main/README.md"
Issues = "https://github.com/cfe-lab/hla_algorithm/issues"
Source = "https://github.com/cfe-lab/hla_algorithm"

[project.scripts]
interpret_from_json = "hla_algorithm.interpret_from_json:main"
bblab = "hla_algorithm.bblab:main"
update_alleles = "hla_algorithm.update_alleles:main"
update_frequency_file = "hla_algorithm.update_frequency_file:main"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.hatch.build]
include = [
  "src/hla_algorithm/_version.py",
  "src/hla_algorithm/__init__.py",
  "src/hla_algorithm/__main__.py",
  "src/hla_algorithm/hla_algorithm.py",
  "src/hla_algorithm/interpret_from_json_lib.py",
  "src/hla_algorithm/interpret_from_json.py",
  "src/hla_algorithm/models.py",
  "src/hla_algorithm/py.typed",
  "src/hla_algorithm/update_alleles.py",
  "src/hla_algorithm/update_frequency_file_lib.py",
  "src/hla_algorithm/update_frequency_file.py",
  "src/hla_algorithm/utils.py",
  "src/hla_algorithm/default_data/hla_standards.yaml",
  "src/hla_algorithm/default_data/hla_frequencies.csv",
]
exclude = [
  "tests",
]
skip-excluded-dirs = true
directory = "output"

[tool.hatch.build.targets.wheel]
packages = ["src/hla_algorithm"]

[tool.hatch.build.hooks.version]
path = "src/hla_algorithm/_version.py"
template = '''
__version__ = "{version}"
'''

[tool.uv]
package = true

[tool.uv-dynamic-versioning]
vcs = "git"
style = "semver"
fallback-version = "0.0.0"

[tool.pytest.ini_options]
pythonpath = "src"
minversion = "6.0"
addopts = "-ra -n 4 --cov --cov-report term --cov-report xml --strict-markers -m 'not slow'"
testpaths = ["tests"]
markers = [
  "integration: Run integration tests",
  "slow: Run integration tests that take a really long time",
]

[tool.coverage.run]
branch = true
parallel = true
source = ["src"]
omit = [
  "src/hla_algorithm/__init__.py",
  "src/hla_algorithm/_version.py",
  "src/hla_algorithm/__main__.py",
  "tests/__init__.py",
  "src/hla_algorithm/bblab.py",
  "src/hla_algorithm/interpret_from_json.py",
  "src/hla_algorithm/update_alleles.py",
  "src/hla_algorithm/update_frequency_file.py",
  "src/scripts/*.py",
]

[tool.coverage.report]
exclude_lines = ["no cov", "if __name__ == .__main__.:", "if TYPE_CHECKING:"]
skip_empty = true

[tool.coverage.xml]
output = 'coverage.xml'

[tool.ruff]
target-version = "py312"

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "C"]
# ignore warnings where `== True` occurs, because we're using a lot of pandas
# https://www.flake8rules.com/rules/E712.html
ignore = ["E712"]


[tool.pydocstyle]
match = "src/**/*.py"

[tool.mypy]
plugins = ["numpy.typing.mypy_plugin"]
ignore_missing_imports = true
exclude = ["scripts/"]
